casts: the act-registrar "free void()" placeholders were member fns - 433 -> 374
Two seams, both the same shape the brief predicted: an apparently API-forced cast
was a wrongly-declared neighbour.

1. SDK constants + typed out-params (-37, byte-neutral)
   - reinterpret_cast<HKEY>(0x80000002) etc -> HKEY_LOCAL_MACHINE / HKEY_CLASSES_ROOT /
     INVALID_HANDLE_VALUE / IDC_ARROW / RT_BITMAP (the SDK already spells them).
   - (LPPOINT)&rc / +1 -> CRect::TopLeft() / BottomRight() (MFC's own accessor pair).
   - The DirectSound / DirectDraw / registry out-params are DWORD (unsigned long),
     not u32 (unsigned int) - that int-vs-long mismatch was the whole reason for
     13 LPDWORD puns. Typed at the declaration, the casts fall out.
   - CNetMgr::SendEx is a pass-through: it now carries IDirectPlay4::SendEx's own
     LPVOID/LPDWORD parameter types.

2. The CActReg act registrars (-22) - the real find
   Every `*(void**)Table.ResolveEntry(id) = (void*)&SomeActHandler` was storing an
   `extern "C" void SomeActHandler()` ILT-thunk placeholder into a pointer-to-MEMBER
   slot. Reading each thunk out of the retail ILT band names the real method:
     ILT 0x4025db -> 0x0b1af0  CSpotLight::Tick          ILT 0x402414 -> 0x0b1ee0  CSpotLight::Update
     ILT 0x40187a -> 0x0419e0  CTeleporter::Begin        ILT 0x403846 -> 0x041aa0  CTeleporter::Update
     ILT 0x403418 -> 0x040d20  CGruntPuddle::Remove      ILT 0x40180c -> 0x0b2ca0  CKitchenSlime::Tick
     ILT 0x403517 -> 0x008440  CActionArea::Tick         ILT 0x403896 -> 0x0e05e0  CProjectile::DetachRenderObj
     ILT 0x403c10 -> 0x0ae2a0  CMenuSparkle::AdvanceAnim ILT 0x402dd8 -> 0x11a8e0  CGruntVoice::Update
   so the store is just `*slot = static_cast<CActHandler>(&C::M)` - the spelling
   Wormhole.cpp already used for CWormhole::SpawnPartners. 14 extern "C" hex
   placeholders retired with them.

   Six thunks pointed at bodies nobody had reconstructed; all six are now in, all
   six at 100%:
     0x11a8c0 CGruntVoice::IdleHidden        0x040c10 CGruntPuddle::Idle
     0x03e500 CGruntStartingPoint::Idle      0x03a910 CCursorSnapSprite::AdvanceAnim
     0x0b5070 CPathHazard::ForwardTick (was mis-declared void - it tail-jumps a
              slot that returns i32)        0x0b5080 CPathHazard::ForwardSiblingTick

   IDENTITY: RbEffect was CExplosion. FortressFlag's RegisterXLogic_6447f8 (0x474b0)
   stores ILT 0x4041ec -> 0x476b0 into CActRegPool<CExplosion>::s_table; an act table
   only ever holds its own class's CUserLogic member fn, and the placeholder's +0x10 /
   +0x38 are exactly CUserLogic::m_object / m_38. RbEffect.h deleted.

3. CActionOptionsMenuBar's six button scalars are three 2-element ARRAYS - proven by
   Serialize's `ar->Write(&m_button0State, 8)` / `(&m_button0Icon, 8)`. Typed as
   arrays, the i32-cursor walk (p[-4]/p[-2]/*p) becomes an indexed loop and the three
   CImage*-through-an-i32-slot puns go.

4. SETTLED, not folded: CGruntSpawnConfig::SpawnVoiceDriver's first arg is an object
   ID. 0x11ba7a/0x11ba82 hand it to CGruntVoice::Setup -> m_source, which Update uses
   as the m_map48 key (the map WwdObjMgr fills with `m_map48[(void*)obj->m_188]`).
   CWarlord's m_object->m_188 is right; the entrance path's `push esi` (0x63073) is
   retail widening a raw CGrunt* into that id slot. Reason replaced, decl named.

Overall 3122/4263 -> 3127/4268 exact, fuzzy 78.57% unchanged, MAX gate holds.

+++ include/Dsndmgr/DirectSoundMgr.h
@@ -40,10 +40,13 @@ public:
     i32 SetFrequency(u32 freq); // 0x135880  SetFrequency (caps DSBCAPS_CTRLFREQUENCY)
     i32 SetField2(i32 pct);     // 0x135920  freq-percent + duration recompute
     void ComputeDuration();     // 0x1359a0  m_durationMs = m_sampleCount*1000/m_sampleRate
-    i32 Unlock(void* p1, u32 n1, void* p2, u32 n2);         // 0x1359c0
-    i32 GetCurrentPosition(u32* play, u32* write);          // 0x135a20
-    i32 SetCurrentPosition(u32 pos);                        // 0x135a70
-    i32 GetFormat(void* fmt, u32 size, u32* written);       // 0x135ac0
+    i32 Unlock(void* p1, u32 n1, void* p2, u32 n2); // 0x1359c0
+    // The DirectSound out-params are DWORD (`unsigned long`), not `unsigned int`:
+    // spelt out so &n binds to LPDWORD directly, and so this header stays Win32-free
+    // (same device as m_pan/m_volume below).
+    i32 GetCurrentPosition(unsigned long* play, unsigned long* write); // 0x135a20
+    i32 SetCurrentPosition(u32 pos);                                   // 0x135a70
+    i32 GetFormat(void* fmt, u32 size, unsigned long* written);        // 0x135ac0
     i32 LoadFromFile(FILE* fp, u32 bytes, i32 offset);      // 0x135e10  fseek+Lock+fread+Unlock
     i32 LockConvert(void* src, u32 lockBytes, u32 convert); // 0x135f40
     i32 Play();                                             // 0x136270  Play + reacquire-retry
@@ -52,9 +55,9 @@ public:
         u32 off,
         u32 bytes,
         void** p1,
-        u32* n1,
+        unsigned long* n1,
         void** p2,
-        u32* n2,
+        unsigned long* n2,
         u32 flags
     ); // 0x136370  Lock + reacquire-on-DSERR_BUFFERLOST retry
 
@@ -70,7 +73,7 @@ public:
     IDirectSoundBuffer* m_buffer; // +0x0c  the held sound buffer
     SoundDevice* m_owner;         // +0x10  owning device back-pointer
     u32 m_playFlags;              // +0x14  Play/looping flags (bit 0 = loop)
-    u32 m_freq;                   // +0x18  cached frequency (GetFrequency)
+    unsigned long m_freq;         // +0x18  cached frequency (DWORD - GetFrequency writes it)
     long m_pan;                   // +0x1c  cached pan (LONG, the type GetPan writes;
                                   //        spelt `long` so this header stays Win32-free)
     long m_volume;                // +0x20  cached volume (LONG, GetVolume writes it)
